chore: rename directory zui-to-typescript-next to zui-to-typescript-type
